[UKRIW-612] switch to new BE API v1.2 (#140)
