+++ apps/storybook-host/.storybook/config/env.config.ts
@@ -62,7 +62,7 @@ const removeTrailingSlashes = (url: string) => {
 };
 
 const getEodhProUrl = () => {
-  const apiVersion = '/v1.1';
+  const apiVersion = '/v1.2';
   const importedUrl = getValue<string>(import.meta.env.VITE_EODH_PRO_API_URL, config?.authorization.clientId, '');
 
   if (importedUrl) {
