+++ libs/map/data-access-map/src/index.ts
@@ -11,7 +11,7 @@ export {
   useFootprintCollectionMutation,
 } from './lib/store/footprint/footprint.store';
 export { useDataSets } from './lib/store/data-sets/data-sets.store';
-export { useGetPresets } from './lib/query/presets.query';
+export { useGetPresets } from './lib/query/presets/presets.query';
 export { useGetHistory } from './lib/query/history/history.query';
 export { useWorkflowStatus } from './lib/query/history/workflow-status.query';
 export type { IHistoryParams } from './lib/query/history/history.query';
@@ -33,11 +33,11 @@ export type {
 } from './lib/store/action-creator/action-creator.model';
 export type { TDataSetValue } from './lib/store/action-creator/action-creator.schema';
 export type { TBaseFunction } from './lib/store/action-creator/node.utils';
-export { useFunctions } from './lib/query/function.query';
+export { useFunctions } from './lib/query/function/function.query';
 export { useCreateWorkflow, useCreateWorkflowStatus } from './lib/mutation/workflow.mutation';
-export type { TFunction } from './lib/query/function.model';
-export type { TPreset } from './lib/query/presets.model';
-export { useCollectionInfo } from './lib/query/collection-info.query';
+export type { TFunction } from './lib/query/function/function.model';
+export type { TPreset } from './lib/query/presets/presets.model';
+export { useCollectionInfo } from './lib/query/collection-info/collection-info.query';
 export { fetchImage } from './lib/query/protected-image.query';
 export type { TMode } from './lib/store/mode.model';
 export { useWorkflow } from './lib/store/workflow/workflow.store';
