+++ libs/map/data-access-map/src/lib/mutation/workflow.utils.ts
@@ -1,7 +1,7 @@
 import { createIsoStringDate } from '@ukri/shared/utils/date';
 import z from 'zod';
 
-import { functionSchema, TFunction } from '../query/function.model';
+import { functionSchema, TFunction } from '../query/function/function.model';
 import { TAreaNode, TDataSetsNode, TDateRangeNode, TFunctionNode } from '../store/action-creator/action-creator.model';
 import { createGeometry } from '../store/results/geometry';
 import { getIntersects, TGeometry } from './get-intersects';
