+++ libs/map/data-access-map/src/lib/query/collection-info/collection-info.model.ts
