+++ libs/map/data-access-map/src/lib/query/presets/presets.query.tsx
@@ -1,9 +1,9 @@
 import { useQuery } from '@tanstack/react-query';
 import { getHttpClient, TExtractFnReturnType } from '@ukri/shared/utils/react-query';
 
-import { paths } from '../api';
+import { paths } from '../../api';
+import { queryKey } from '../query-key.const';
 import { presetsSchema, TPreset } from './presets.model';
-import { queryKey } from './query-key.const';
 
 const presets = async (): Promise<TPreset[]> => {
   const response = await getHttpClient().get(paths.PRESETS);
